Refactoring code of cards on the Identities page
