--- packages/frontend/src/components/identities/TopIdentities.scss
@@ -1,3 +1,0 @@
-.TopIdentities {
-    
-}
